# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Math Functions API
  version: 1.0.0
  description: API для работы с математическими функциями
servers:
  - url: http://localhost:8080/api
    description: Local server
security:
  - basicAuth: []
paths:
...
  /access:
    get:
      summary: Получить список пользователей с доступом к функции
      tags: [Access]
      parameters:
        - name: functionId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessResponse'
    post:
      summary: Предоставить доступ пользователю к функции
      tags: [Access]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                functionId:
                  type: integer
                userId:
                  type: integer
                accessType:
                  type: string
                  enum: [READ, WRITE]
      responses:
        '201':
          description: Доступ предоставлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessResponse'
    delete:
      summary: Отозвать доступ
      tags: [Access]
      parameters:
        - name: functionId
          in: query
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Доступ отозван

  # МАТЕМАТИЧЕСКИЕ ОПЕРАЦИИ
...
components:
  schemas:
...
    AccessResponse:
      type: object
      properties:
        accessId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        accessType:
          type: string
        grantedAt:
          type: string
          format: date-time
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic authentication with username and password